// packages/apps/human-app/server/src/common/enums/endpoint-name.ts
export enum EndpointName {
  WORKER_SIGNUP = 'WORKER_SIGNUP',
  OPERATOR_SIGNUP = 'OPERATOR_SIGNUP',
  WORKER_SIGNIN = 'WORKER_SIGNIN',
  EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
  RESEND_EMAIL_VERIFICATION = 'RESEND_EMAIL_VERIFICATION',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  RESTORE_PASSWORD = 'RESTORE_PASSWORD',
  PREPARE_SIGNATURE = 'PREPARE_SIGNATURE',
  DISABLE_OPERATOR = 'DISABLE_OPERATOR',
}